ai-service/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Computer Vision & AI (Optimized)
opencv-python-headless==4.8.1.78
numpy==1.24.3
insightface==0.7.3
onnxruntime-gpu==1.17.1  # Prefer GPU if available, falls back to CPU cleanly usually
onnxruntime==1.17.1      # CPU fallback explicitly just in case

# Face Enhancement (GFPGAN - Optional but powerful)
# Kept because user has powerful hardware
gfpgan==1.3.8
realesrgan==0.3.0
facexlib==0.3.0
basicsr==1.4.2
torch==2.1.0
torchvision==0.16.0

# Database
aiomysql==0.2.0
PyMySQL==1.1.0

# Object Storage
minio==7.2.0

# Image Processing
Pillow==10.1.0
rembg==2.0.52

# Utilities
requests==2.31.0
python-dotenv==1.0.0